% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDQ.R
\name{deepDQ}
\alias{deepDQ}
\title{Deep analysis of data quality}
\usage{
deepDQ(
  data,
  metric,
  columnDate = NULL,
  var_time_name = NULL,
  position = FALSE,
  dataref = NULL,
  ranges = NULL,
  maxdif = NULL,
  units = "secs"
)
}
\arguments{
\item{data}{The data frame or time series to be analyzed}

\item{metric}{The name of the metric to be analyzed}

\item{var_time_name}{The name of the variable containing the date or time information}

\item{position}{Logical argument. If TRUE date or time information is added to the results}

\item{dataref}{A data frame with the correct format to compare the current data}

\item{ranges}{A data frame containing the minimum and maximum value allowed to each variable}
}
\value{
Detailed information about the failures occurred in the chosen metric
}
\description{
The data set is entered and the name of the metric to be analyzed in depth is indicated. Data quality can be evaluated by variables or the values in which an error occurs can be shown.
}
