% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dqplot}
\alias{dqplot}
\title{Plot data quality metrics}
\usage{
dqplot(DQ, totalquality = FALSE)
}
\arguments{
\item{DQ}{A data frame with quality metrics as columns}

\item{totalquality}{Logical argument. If TRUE the global data quality is added with the other metrics}
}
\value{
Plot of the quality metric values
}
\description{
Takes a data quality data frame and plot the values of the metrics
}
