% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandlingLowQuality.R
\name{handleDQ}
\alias{handleDQ}
\title{Handle low data quality}
\usage{
handleDQ(
  data,
  metric,
  columnDate = NULL,
  var_time_name = NULL,
  ranges = NULL,
  dataref = NULL,
  method = "mean",
  maxdif = NULL,
  units = NULL
)
}
\arguments{
\item{data}{The data frame with low data quality to be corrected}

\item{metric}{the name of the metric to be corrected}

\item{var_time_name}{optional name of the time variable}

\item{ranges}{A data frame containing the minimum and maximum value allowed to each variable}
}
\value{
the data frame with the low quality values corrected
}
\description{
This function allows to correct the low data quality of one of the metrics. Once the process is done, it returns the data frame with some values modified. Do, the new value for the metric chosen is 1.
}
