% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.quality}
\alias{plot.quality}
\title{Plot data quality metrics}
\usage{
\method{plot}{quality}(DQ, totalquality = FALSE)
}
\arguments{
\item{DQ}{A data frame with quality metrics as columns}
}
\value{
Plot of the quality metric values
}
\description{
Takes a data quality fata frame and plot the values of the metrics
}
