% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{DQ}
\alias{DQ}
\title{Data quality from a data set}
\usage{
DQ(
  data,
  columnDate,
  maxdif,
  dataref,
  ranges = NULL,
  weights = NULL,
  windows = FALSE,
  cte = TRUE,
  fixed = TRUE,
  nint = 3,
  by = 10
)
}
\arguments{
\item{data}{A data frame to be analyzed}
}
\value{
A data frame containing the values of all quality metrics in columns
}
\description{
Takes a data set and the conditions to compute the data quality and returns the values of all metrics
}
